Move InboxState to new file.
